requirements.txt
numpy
opencv-python
matplotlib
scikit-learn
Pillow
torch